rename i18n file and references to NumberFormat
